requirements.txt
openai==0.28.1
google-genai==1.23.0
redis==4.6.0
langgraph
httpx
pytest
python-dotenv
